% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_read_sharepoint.R
\name{ers_read_sharepoint}
\alias{ers_read_sharepoint}
\title{Read files from internal sharepoint folders}
\usage{
ers_read_sharepoint(
  folder_path,
  file_name_with_extension,
  sheet_name = NULL,
  drive_name = "client_work_drive",
  skip_rows = 0,
  na_values = c("", "N/A", "NA", "Missing")
)
}
\arguments{
\item{folder_path}{REQUIRED: The path to the file. This can be pasted directly from SharePoint: https://app.tettra.co/teams/ersknowledge/pages/copying-a-folder-path-from-sharepoint-to-use-in-r}

\item{file_name_with_extension}{REQUIRED: File name. Example: "student_performance.csv"}

\item{sheet_name}{OPTIONAL: Only required if you have an excel spreadsheet with multiple sheets. Example: "Sheet 2"}

\item{drive_name}{OPTIONAL: R will detect what drive the file is from (e.g. 'client_work_drive', 'internal_drive', 'external_drive', or 'data_hub_drive')}

\item{skip_rows}{OPTIONAL: Only required if your data doesn't start in the first row of the spreadsheet. Example: If your data starts in row 4 then enter 3}

\item{na_values}{OPTIONAL: By default will change all cells to NA if they have "", "N/A", "NA", or "Missing"}
}
\value{
Your loaded data as a data frame with the name raw_data
}
\description{
Read files from internal sharepoint folders
}
