% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_sharepoint_path_clean.R
\name{ers_sharepoint_path_clean}
\alias{ers_sharepoint_path_clean}
\title{Clean beginning path and \%20s in sharepoint file names for ers_read_sharepoint and ers_write_sharpoint}
\usage{
ers_sharepoint_path_clean(full_path)
}
\arguments{
\item{full_path}{REQUIRED: The path to the file. This can be pasted directly from SharePoint: https://app.tettra.co/teams/ersknowledge/pages/copying-a-folder-path-from-sharepoint-to-use-in-r}
}
\value{
Your cleaned sharepoint path and the sharepoint drive name
}
\description{
Clean beginning path and \%20s in sharepoint file names for ers_read_sharepoint and ers_write_sharpoint
}
