% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_template.R
\name{ers_template}
\alias{ers_template}
\title{Create a new ERS R Markdown file from a template}
\usage{
ers_template(file_name, include_walkthrough = FALSE)
}
\arguments{
\item{file_name}{The name of the new Rmd file to create (e.g., "my_report"). You do not need to include the .Rmd extension.}

\item{include_walkthrough}{If TRUE or "yes", use the walkthrough template; otherwise, use the standard template}
}
\value{
Creates a new Rmd file in your working directory
}
\description{
Create a new ERS R Markdown file from a template
}
