% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_write_sharepoint.R
\name{ers_write_sharepoint}
\alias{ers_write_sharepoint}
\title{Write files to internal sharepoint folders}
\usage{
ers_write_sharepoint(
  data,
  folder_path,
  file_name_with_extension,
  drive_name = "client_work_drive"
)
}
\arguments{
\item{data}{REQUIRED: The data frame to be uploaded}

\item{folder_path}{REQUIRED: The folder where you want the data to be saved. Example: "District Partners/Yakima". Folder path must start with either "District Partners" or "Internal"}

\item{file_name_with_extension}{REQUIRED: File name. Example: "student_performance.csv"}

\item{drive_name}{OPTIONAL: By default it is set to client work, if you want to access something from internal, then change to internal_drive}
}
\value{
A message that your file was successfully uploaded.
}
\description{
Write files to internal sharepoint folders
}
