% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sharepoint_ers.R
\name{read_sharepoint_ers}
\alias{read_sharepoint_ers}
\title{Read files from internal sharepoint folders}
\usage{
read_sharepoint_ers(
  folder_path,
  file_name_with_extension,
  drive_name = client_work_drive,
  sheet_name = NULL,
  skip_rows = 0,
  na_values = c("", "N/A", "NA", "Missing")
)
}
\arguments{
\item{folder_path}{REQUIRED: The path to the file. Example: "District Partners/Yakima". Folder path must start with either "District Partners" or "Internal"}

\item{file_name_with_extension}{REQUIRED: File name. Example: "student_performance.csv"}

\item{drive_name}{OPTIONAL: By default it is set to client work, if you want to access something from internal, then change to internal_drive}

\item{sheet_name}{OPTIONAL: Only required if you have an excel spreadsheet with multiple sheets. Example: "Sheet 2"}

\item{skip_rows}{OPTIONAL: Only required if your data doesn't start in the first row of the spreadsheet. Example: If your data starts in row 4 then enter 3}

\item{na_values}{OPTIONAL: By default will change all cells to NA if they have "", "N/A", "NA", or "Missing"}
}
\value{
Your loaded data as a data frame with the name raw_data
}
\description{
Read files from internal sharepoint folders
}
