% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_column_info.R
\name{ers_column_info}
\alias{ers_column_info}
\title{Use the skim function to validate your data and output a summary of your columns}
\usage{
ers_column_info(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble to be summarized.}
}
\value{
A tibble with summary statistics, where numeric values are rounded to one decimal place
and large numbers (≥ 1,000) are formatted with commas.
}
\description{
This function applies \code{skimr::skim()} to summarize a dataset and then formats numeric columns.
Numeric values are rounded to one decimal place. If a number is 1,000 or greater, it is formatted with commas.
}
