% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_update_erstools.R
\name{ers_update_erstools}
\alias{ers_update_erstools}
\title{Update the erstools package from GitHub}
\usage{
ers_update_erstools()
}
\value{
Invisibly returns TRUE if installation succeeds.
}
\description{
Installs or updates the public erstools package from GitHub.
}
