% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ers_write_sharepoint.R
\name{ers_write_sharepoint}
\alias{ers_write_sharepoint}
\title{Write files to internal sharepoint folders}
\usage{
ers_write_sharepoint(
  data,
  folder_path,
  file_name_with_extension,
  drive_name = "client_work_drive"
)
}
\arguments{
\item{data}{REQUIRED: The data frame, shape file, or ggplot object to be uploaded}

\item{folder_path}{REQUIRED: The folder where you want the data to be saved: https://app.tettra.co/teams/ersknowledge/pages/copying-a-folder-path-from-sharepoint-to-use-in-r}

\item{file_name_with_extension}{REQUIRED: File name. Example: "student_performance.csv"}

\item{drive_name}{OPTIONAL: R will detect what drive the file is from (e.g. 'client_work_drive', or 'internal_drive')}
}
\value{
A message that your file was successfully uploaded.
}
\description{
Write files to internal sharepoint folders
}
